Refactor out resource test base
